// src/models/saksoversikt/fellesEnum.ts
export enum Baksystem {
    Gsak = 'GSAK',
    Pesys  = 'PESYS',
    SakOgBehandling = 'SAK_OG_BEHANDLING',
    Joark = 'JOARK',
    JoarkSikkerhetsbegrensning = 'JOARK_SIKKERHETSBEGRENSNING',
    Henvendelse = 'HENVENDELSE',
    PdfKonvertering = 'PDF_KONVERTERING',
    Aktoer = 'AKTOER',
    Kodeverk = 'KODEVERK'
}